% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_agreement.R
\name{calculate_agreement}
\alias{calculate_agreement}
\title{Calculate Agreement Rates for Annotation Values}
\usage{
calculate_agreement(df, nmm_values)
}
\arguments{
\item{df}{A confusion matrix, without a row/column with total scores}

\item{nmm_values}{A list of the annotation values for which agreement rates should be calculated.}
}
\value{
agreement: A dataframe with scores for raw agreement, chance agreement, and Cohen's kappa for each annotation value.
}
\description{
This function calculates the agreement rates for the specified annotation values.
}
