% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_kappa.R
\name{cohens_kappa}
\alias{cohens_kappa}
\title{Calculate Cohen's Kappa}
\usage{
cohens_kappa(df, cat)
}
\arguments{
\item{df}{A confusion matrix (without a row and column for "total")}

\item{cat}{A string specifying a category.}
}
\value{
p_o: The raw agreement score

p_e: The expected agreement score

kappa: The Cohen's Kappa score

collapsed:  The collapsed confusion matrix
}
\description{
This function calculates the raw agreement, expected agreement, and Cohen's Kappa for a given category.
}
