% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Create Event-based Confusion Matrix}
\usage{
confusion.matrix(df)
}
\arguments{
\item{df}{Dataframe as processed and prepared by the match function.}
}
\value{
confusion: The confusion matrix for the event-based approach

confusion.total: The confusion matrix for the event-based approach, with a 'total' row and column.
}
\description{
This function creates the event-based confusion matrix. The rows in the matrix represent Annotator.a, and the columns represent Annotator.b.
}
