% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.drop.levels.R
\name{filter.drop.levels}
\alias{filter.drop.levels}
\title{Filter dataframe and drop levels}
\usage{
\method{filter}{drop.levels}(df, condition)
}
\arguments{
\item{df}{A dataframe}

\item{condition}{A string specifying a condition.}
}
\value{
df.filtered:  The filtered dataframe.
}
\description{
This function filters a dataframe on a given condition, and drops the levels in factor columns.
}
