% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{matches}
\alias{matches}
\title{Prepare Dataframe for Event-based IRR Calculations}
\usage{
matches(df)
}
\arguments{
\item{df}{The pre-processed dataframe.}
}
\value{
df.event.matches: The dataframe that will be used to create the confusion matrix.
}
\description{
This function prepares the dataframe for event-based IRR calculations. In this case, we remove the "neutral" annotations. Any annotations that do not have at least the specified overlap threshold are labeled as "unmatched", and are commission/omission errors.
}
