% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Calculate overlap between annotations}
\usage{
overlap(df)
}
\arguments{
\item{df}{The pre-processed dataframe.}
}
\value{
overlapping: A dataframe denoting the percentage overlap of between annotations.
}
\description{
This function calculates the overlap between annotations for each question. Overlap is calculated as the percentage of overlapping frames relative to the annotation with the highest number of frames.
}
