% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.agreement.R
\name{plot.agreement}
\alias{plot.agreement}
\title{Plot Agreement Rates}
\usage{
\method{plot}{agreement}(agreement, save.plot = FALSE)
}
\arguments{
\item{agreement}{A dataframe with agreement scores}

\item{save.plot}{A boolean specifiying whether or not the plot should be saved}

\item{frame}{A boolean specifying if the method is frame-based or not. Should be removed.}
}
\value{
A bar graph of the agreement rates for each annotation value. This is either saved to a folder or shown.
}
\description{
This function plots the agreement rates from an agreement dataframe.
}
