% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.matched.annotations.R
\name{plot.matched.annotations}
\alias{plot.matched.annotations}
\title{Plot Annotations}
\usage{
\method{plot}{matched.annotations}(df, question, threshold = FALSE)
}
\arguments{
\item{df}{The pre-processed dataframe}

\item{question}{A string of a question ID}

\item{threshold}{A boolean indicating whether matches below the overlap threshold should be dropped.}
}
\value{
p: A plot of the annotations by both annotators for the question. The plot includes a line between overlapping annotations and indicates the percentage overlap between them.
}
\description{
This function plots annotations for a given question.
}
