% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize Dataframe for given tier}
\usage{
init(tier, replace_value = FALSE)
}
\arguments{
\item{tier}{A string of desired tier name. Possible tiers: body_position, eye_gaze, eye_shape (closed), eyebrows, head_move, head_x, head_y, head_z, lip_corners (mouth.action), lips (mouth.action), nose, shoulders}

\item{replace_value}{A boolean, whether values need to be replaced}
}
\value{
df: A pre-processed dataframe for the given tier
}
\description{
This function initializes the dataframe for a given tier. It pre-processes the data into a standardized form.
}
