% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{matches}
\alias{matches}
\title{Prepare Dataframe for Event-based IRR Calculations}
\usage{
matches(df, Annotator.a = "C1", Annotator.b = "C2", Overlap.threshold = 51)
}
\arguments{
\item{df}{The pre-processed dataframe.}

\item{Annotator.a}{Annotator label A - name of annotator}

\item{Annotator.b}{Annotator label B - name of annotator}

\item{Overlap.threshold}{overlapping threshold value (default = 51)}
}
\value{
df.event.matches: The dataframe that will be used to create the confusion matrix.
}
\description{
This function prepares the dataframe for event-based IRR calculations. In this case, we remove the "neutral" annotations. Any annotations that do not have at least the specified overlap threshold are labeled as "unmatched", and are commission/omission errors.
}
